Fix raw to voltage conversion math
Single-ended measurements taken by this demo were off by a factor of two, due to how the conversion was performed - the conversion originally used a division by 0xFFFF to convert data from a 16-bit signed integer to a float (which would have been correct for an unsigned integer). Additionally, the differential channel scale was set to 1.0 when measurements range from -0.5 to 0.5V, since the raw data is signed, this has been adjusted to a scale of 0.5.

Differential measurements were reported correctly in the original version, but were arrived at in the wrong way. by correcting the resolution for the single-ended measurements, they would break without the adjustment to the scale.