Fix: single-ended analog input channels
Changed the input mode for single-ended analog input channels to unipolar.

Added conversion for single-ended channels.

Modified full scale code to 0x7FF0 for differential channels and to 0xFFF0 for single-ended channels, because the 12-bit conversion data correspond to the 12 MSBs in the 16-bit status registers according to the Xilinx XADC User Guide (UG480).